Removed QvS function from FMMM (QvS just overwrote iteration settings)

+++ src/vecops.h
@@ -3,5 +3,5 @@
 #include <ogdf/basic/GraphAttributes.h>
 #include <QPointF>
 
-float length(const QPointF& p);
+double length(const QPointF& p);
 QPointF normalize(const QPointF& p);
